fix(ci): python-full-coverage — split execution_validation fix-pass loop into focused module
