+++ tests/unit/control/test_executor_coverage_edges_parts/test_executor_coverage_edges_part_007.py
@@ -13,6 +13,7 @@ from awf.adapters.base import AgentRunError
 from awf.common.commands import CommandResult
 from awf.common.compose_exec import ComposeExecCleanupError
 from awf.control.executor import execution_validation as executor_execution_validation
+from awf.control.executor import hosted_validation_sync as executor_hosted_validation_sync
 from awf.control.executor import validation_cleanup_guards as executor_validation_cleanup_guards
 from awf.control.executor import validation_fix_helpers as executor_validation_fix_helpers
 from awf.control.executor.types import _RebaseRecoveryResult
@@ -976,7 +977,7 @@ async def test_hosted_validation_fix_head_sync_reports_failures(
 ) -> None:
     """Hosted fix head sync failures should keep actionable reason codes."""
     runner = SimpleNamespace(run=AsyncMock(side_effect=queued_results))
-    result = await executor_execution_validation._sync_hosted_validation_fix_head(  # noqa: SLF001
+    result = await executor_hosted_validation_sync._sync_hosted_validation_fix_head(  # noqa: SLF001
         SimpleNamespace(_runner=runner),
         worktree_path=tmp_path / "worktree",
         hosted_pr_identity=identity,
@@ -1003,7 +1004,7 @@ def test_hosted_agent_error_terminal_head_prefers_result_sha_over_details() -> N
         details={"terminal_head_sha": "e" * 40},
     )
 
-    assert executor_execution_validation._hosted_agent_error_terminal_head_sha(exc) == "d" * 40
+    assert executor_hosted_validation_sync._hosted_agent_error_terminal_head_sha(exc) == "d" * 40
 
 
 @pytest.mark.unit
